rename common.py to analysis.py
